+++ boostmtree/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(boostmtree)
+
+test_check("boostmtree")